+++ NEWS.md
@@ -1,3 +1,21 @@
+# mixAR 0.22.6 (CRAN)
+
+- `fit_mixAR` was failing for some mixAR models with Gaussian components when
+  the AR order of a component was zero. (reported by Lu Cao)
+
+- improved printing of `mixARgen` objects.
+
+- The function returned by the methods for `multiStep_dist` with `N = "missing"`
+  now checks the length of the supplied `xcond` argument and throws error if it
+  is shorter than the maximal AR order `p`. If it is longer, the last `p` values
+  in it are used.  This has always been the case when `N` is not missing.
+
+- new function `mix_qf` computes conditional quantiles, analogous to `mix_pdf`
+  and `mix_cdf`.
+
+- numerous fixes and improvements.
+
+
 # mixAR 0.22.5 (CRAN)
 
 - in tests, include `check.environment = FALSE` in calls to expect_equal() to